# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Booking for All API",
                 "description":  "Comprehensive appointment management system API. Connect organizations, service providers, and clients seamlessly.",
                 "version":  "1.0.0",
                 "contact":  {
                                 "name":  "API Support"
                             }
             },
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT",
                                                                  "description":  "Bearer token authentication. Include the token in the Authorization header as: Bearer {token}"
                                                              },
                                               "apiKey":  {
                                                              "type":  "apiKey",
                                                              "in":  "header",
                                                              "name":  "x-api-key",
                                                              "description":  "API key authentication for external integrations"
                                                          }
                                           },
...
                   },
    "paths":  {
...
                  "/api/provider/events/{id}":  {
                                                    "put":  {
                                                                "parameters":  [
                                                                                   {
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                       "in":  "path",
                                                                                       "name":  "id",
                                                                                       "required":  true
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Default Response"
                                                                                          }
                                                                              }
                                                            },
                                                    "delete":  {
                                                                   "parameters":  [
                                                                                      {
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     },
                                                                                          "in":  "path",
                                                                                          "name":  "id",
                                                                                          "required":  true
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Default Response"
                                                                                             }
                                                                                 }
                                                               }
                                                },
...
                  "/api/events/{id}":  {
                                           "get":  {
                                                       "parameters":  [
                                                                          {
                                                                              "schema":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                              "in":  "path",
                                                                              "name":  "id",
                                                                              "required":  true
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Default Response"
                                                                                 }
                                                                     }
                                                   }
                                       },
...
              },
    "servers":  [
                    {
                        "url":  "http://localhost:3000",
                        "description":  "API Server"
                    }
                ],
...
}